added authorization

+++ backend/routes/posts.js
@@ -40,7 +40,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
     });
 
     post.save().then(createdPost => {
@@ -69,11 +70,17 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: 'Update successful!' });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0) {
+        res.status(200).json({ message: 'Update successful!' });
+      }
+      else {
+        res.status(401).json({ message: 'Not Authorised!' });
+      }
     });
   }
 );
@@ -111,9 +118,15 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: 'Post deleted!' });
+    if(result.n > 0) {
+      res.status(200).json({ message: 'Post deleted!' });
+    }
+    else {
+      res.status(401).json({ message: 'Not Authorised!' });
+    }
+
   });
 });
